# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistema de Gerenciamento e Criação de Propostas Comerciais
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stations:
    get:
      operationId: listStations
      tags: [stations]
      summary: List stations (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of stations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Station"
    post:
      operationId: createStation
      tags: [stations]
      summary: Create station (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StationInput"
      responses:
        "201":
          description: Station created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"
  /stations/{id}:
    patch:
      operationId: updateStation
      tags: [stations]
      summary: Update station (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StationUpdate"
      responses:
        "200":
          description: Station updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Station"

  # ─── ADVERTISERS ────────────────────────────────────────────────────────────
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List proposals (own for COMERCIAL, all for ADMIN) with pagination and filters
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: advertiserId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated proposals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalListResponse"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a new proposal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Proposal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get proposal full detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proposal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Save proposal changes (auto-creates version)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalUpdate"
      responses:
        "200":
          description: Proposal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Archive/delete proposal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proposal archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /proposals/{id}/status:
    patch:
      operationId: updateProposalStatus
      tags: [proposals]
      summary: Change proposal status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}/duplicate:
    post:
      operationId: duplicateProposal
      tags: [proposals]
      summary: Duplicate a proposal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Proposal duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}/versions:
    get:
      operationId: listProposalVersions
      tags: [proposals]
      summary: List proposal version history
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProposalVersion"
  /proposals/{id}/versions/{versionId}:
    get:
      operationId: getProposalVersion
      tags: [proposals]
      summary: Get a specific version snapshot
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Version snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProposalVersion"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [proposals]
      summary: Dashboard summary stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, COMERCIAL]
        active:
          type: boolean
        createdAt:
          type: string
...
    Station:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slogan:
          type: ["string", "null"]
        logoBase64:
          type: ["string", "null"]
        createdAt:
          type: string
    StationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slogan:
          type: string
        logoBase64:
          type: string
    StationUpdate:
      type: object
      properties:
        name:
          type: string
        slogan:
          type: string
        logoBase64:
          type: string

    # ─── ADVERTISERS ──────────────────────────────────────────────────────────
    Advertiser:
      type: object
      required: [id, tradeName, active, createdAt]
      properties:
        id:
          type: string
        tradeName:
          type: string
        legalName:
          type: ["string", "null"]
        cnpj:
          type: ["string", "null"]
        logoBase64:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
...
    StatBlock:
      type: object
      required: [num, suf, desc]
      properties:
        num:
          type: string
        suf:
          type: string
        desc:
          type: string
    ProposalProduct:
      type: object
      required: [id, order, qty, title, color]
      properties:
        id:
          type: string
        order:
          type: integer
        qty:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        detail:
          type: ["string", "null"]
        program:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        color:
          type: string
          enum: [BLUE, YELLOW, RED, GREEN, DARK]
    Proposal:
      type: object
      required: [id, stationId, createdById, status, propType, propMonth, propYear, stats, overlayOpacity, createdAt, updatedAt]
      properties:
        id:
          type: string
        stationId:
          type: string
        station:
          $ref: "#/components/schemas/Station"
        advertiserId:
          type: ["string", "null"]
        advertiser:
          $ref: "#/components/schemas/Advertiser"
        createdById:
          type: string
        createdBy:
          $ref: "#/components/schemas/User"
        status:
          type: string
          enum: [DRAFT, SENT, APPROVED, REJECTED, ARCHIVED]
        fromTemplateId:
          type: ["string", "null"]
        fromTemplateName:
          type: ["string", "null"]
        propType:
          type: string
        propMonth:
          type: string
        propYear:
          type: string
        campTag:
          type: ["string", "null"]
        clientLine1:
          type: ["string", "null"]
        clientLine2:
          type: ["string", "null"]
        dateStart:
          type: ["string", "null"]
        dateEnd:
          type: ["string", "null"]
        periodDesc:
          type: ["string", "null"]
        bannerBase64:
          type: ["string", "null"]
        overlayOpacity:
          type: integer
        stats:
          type: array
          items:
            $ref: "#/components/schemas/StatBlock"
        investDesc:
          type: ["string", "null"]
        investValue:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactRole:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProposalProduct"
        createdAt:
          type: string
        updatedAt:
          type: string
    ProposalSummary:
      type: object
      required: [id, status, propType, propMonth, propYear, createdAt, updatedAt]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [DRAFT, SENT, APPROVED, REJECTED, ARCHIVED]
        propType:
          type: string
        propMonth:
          type: string
        propYear:
          type: string
        campTag:
          type: ["string", "null"]
        clientLine1:
          type: ["string", "null"]
        advertiserName:
          type: ["string", "null"]
        fromTemplateName:
          type: ["string", "null"]
        createdByName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    ProposalListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProposalSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProposalInput:
      type: object
      required: [propType, propMonth, propYear]
      properties:
        stationId:
          type: string
        advertiserId:
          type: string
        propType:
          type: string
        propMonth:
          type: string
        propYear:
          type: string
        campTag:
          type: string
        clientLine1:
          type: string
        clientLine2:
          type: string
        dateStart:
          type: string
        dateEnd:
          type: string
        periodDesc:
          type: string
        bannerBase64:
          type: string
        overlayOpacity:
          type: integer
        stats:
          type: array
          items:
            $ref: "#/components/schemas/StatBlock"
        investDesc:
          type: string
        investValue:
          type: string
        contactName:
          type: string
        contactRole:
          type: string
        contactPhone:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProposalProductInput"
    ProposalUpdate:
      type: object
      properties:
        advertiserId:
          type: string
        propType:
          type: string
        propMonth:
          type: string
        propYear:
          type: string
        campTag:
          type: string
        clientLine1:
          type: string
        clientLine2:
          type: string
        dateStart:
          type: string
        dateEnd:
          type: string
        periodDesc:
          type: string
        bannerBase64:
          type: string
        overlayOpacity:
          type: integer
        stats:
          type: array
          items:
            $ref: "#/components/schemas/StatBlock"
        investDesc:
          type: string
        investValue:
          type: string
        contactName:
          type: string
        contactRole:
          type: string
        contactPhone:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProposalProductInput"
    ProposalProductInput:
      type: object
      required: [title]
      properties:
        order:
          type: integer
        qty:
          type: string
        title:
          type: string
        description:
          type: string
        detail:
          type: string
        program:
          type: string
        tags:
          type: array
          items:
            type: string
        color:
          type: string
          enum: [BLUE, YELLOW, RED, GREEN, DARK]
    ProposalStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [DRAFT, SENT, APPROVED, REJECTED, ARCHIVED]
    ProposalVersion:
      type: object
      required: [id, proposalId, createdAt]
      properties:
        id:
          type: string
        proposalId:
          type: string
        snapshot:
          type: object
        createdAt:
          type: string
        createdById:
          type: ["string", "null"]

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
    DashboardStats:
      type: object
      required: [total, draft, sent, approved, rejected, archived]
      properties:
        total:
          type: integer
        draft:
          type: integer
        sent:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        archived:
          type: integer
...